ProgramacionIVContenido/tareas/tarea_8_Envio_de_Correos_Asincronos_con_Celery_y_KeyDB/requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
Celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
